added guns and players as prefabs
rifle, pistol and shotgun are prefabs as well as ovr player that can grab them New scene created for testing to show that they work in an entirely new scene. Issue with snap offset. snap offset cannot be a child of the object or it doesn't work. snap offsets are included separately from the prefab