finished holsters and added saddle
added a saddle for the horse, added holsters for the rifle and shotguun as well as fixing holstering isses. created prefabs for the horse and saddle as well as just the saddle. created prefabs for all the holsters and a player with the hip holsters and utility belt attached. also created a utility belt prfefab